Parameterize output filename
Problem: Output filename is constructed from input filename.
In addition to reducing flexibility, the output file is destroyed
in the event of an error.

Solution: Add the ability to, optionally, specify the output filename as
a second argument.  While the file destruction problem is not directly addressed,
that problem can now be solved externally to gitdown.